#define BFRAMES
git-svn-id: http://svn.xvid.org/trunk/xvidcore@152 060a26f8-e6a2-0410-8959-474133d486e0

+++ src/bitstream/mbcoding.c
@@ -511,8 +511,8 @@ void put_bvop_dbquant(Bitstream *bs, int value)
 
 void MBCodingBVOP(const MACROBLOCK * mb, 
 				  const int16_t qcoeff[6*64],
-				  const int16_t fcode,
-				  const int16_t bcode,
+				  const int32_t fcode,
+				  const int32_t bcode,
 				  Bitstream * bs, 
 				  Statistics * pStat)
 {
